requirements.txt
yt-dlp
azure-storage-blob
whisper
ffmpeg-python
soundfile
aiohttp
